Fix Corsair headset stands and mouse pads.
The ApplyMapStandLighting method was checking that a mousemat existed before actually trying to adjust the headset stand. This also fixes having a Corsair mouse pad installed, as it would succeed in the check and then try to adjust the nonexistent headset stand. So, if you had one but not the other, neither worked (and in fact the debug log got spammed from the NRE).